Refactored index.js in main and seperated logic into modular functions inside twilio. Refactored settings.js to reflect changes onto index.
